changing the method of bazaar requests
i have changed the method for requesting item to purchase from bazaar
and use IabHelper classes instead of normal requests.
i made another activity instead of purchase.class and change the intent
for that Activity , ofcourse the purchase.class is exist for research on
it,
i commented some parts of MainPage Activity and Wizard Activity and they
are the previous method for loading purchasable items from bazaar.